YARP_sig: fix warning -Wsign-compare in SoundFile.cpp

+++ src/libYARP_sig/src/SoundFile.cpp
@@ -276,8 +276,8 @@ size_t yarp::sig::file::soundStreamReader::readBlock(Sound& dest, size_t block_s
     dest.setFrequency(soundInfo.freq);
 
     int ct = 0;
-    for (int i=0; i<samples_read; i++) {
-        for (int j=0; j<soundInfo.channels; j++) {
+    for (size_t i=0; i<samples_read; i++) {
+        for (size_t j=0; j< (size_t) soundInfo.channels; j++) {
             dest.set(data[ct],i,j);
             ct++;
         }
